pyproject.toml
[tool.poetry]
name = "cpp-shm-stream"
version = "0.0.0"
description = ""
authors = ["Kenta Kabashima <kenta_program37@hotmail.co.jp>"]
readme = "README.md"
packages = [{ include = "scripts" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
conan = "^1.55.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
breathe = "^4.34.0"
click = "^8.1.3"
cmakelang = "^0.6.13"
doc8 = "^0.11.2"
myst-parser = "^0.18.1"
pre-commit = "^2.20.0"
pylint = "^2.15.8"
sphinx = "^5.3.0"
sphinx-autobuild = "^2021.3.14"
sphinx-rtd-theme = "^1.1.1"
sphinxcontrib-plantuml = "^0.24.1"
clang-tidy-checker = "^0.1.2"
lcov_cobertura = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
disable = "too-many-instance-attributes,line-too-long"
